requirements/requirements.txt
anndata==0.8.0
anndata2ri==1.1
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
contourpy==1.0.5
cryptography==37.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dunamai==1.13.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
fonttools==4.37.4
get_version==3.5.4
gitdb==4.0.9
GitPython==3.1.27
graphviz==0.20.1
h5py==3.7.0
idna==3.3
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.6
jsonschema==4.16.0
jupyter==1.0.0
jupyter_client==7.3.5
jupyter-console==6.4.3
jupyter_core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.1.2
jupyterlab_server==2.15.2
jupyterlab-widgets==1.0.0
kiwisolver==1.4.4
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
natsort==8.2.0
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==2.8.7
notebook==6.4.12
numba==0.56.2
numexpr==2.8.3
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.20
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pycallgraph3==1.1.4
pycparser==2.21
Pygments==2.11.2
pynndescent==0.5.7
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.0
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
rpy2==3.5.4
scanpy==1.8.2
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
Send2Trash==1.8.0
setuptools==63.4.1
sinfo==0.3.4
sip==6.6.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
statsmodels==0.13.2
stdlib-list==0.8.0
tables==3.7.0
terminado==0.13.1
testpath==0.6.0
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
umap-learn==0.3.10
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
widgetsnbextension==3.5.2
